refactor: 불필요 코드 조정, @ControllerTest 커스텀
