+++ Moyoy-Api/build.gradle
@@ -16,7 +16,6 @@ dependencies {
     testImplementation 'org.springframework.security:spring-security-test'
     testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
     testImplementation 'com.epages:restdocs-api-spec-mockmvc:0.19.2'
-    testImplementation 'com.h2database:h2'
 
     // 독립 실행 가능 모듈들은 서로 별도의 로그 포맷 사용
     implementation 'net.logstash.logback:logstash-logback-encoder:7.4'
