+++ Moyoy-Api/src/main/java/com/moyoy/api/auth/jwt/application/response/ReissueJwtResult.java
@@ -1,4 +1,4 @@
-package com.moyoy.api.auth.jwt.application;
+package com.moyoy.api.auth.jwt.application.response;
 
 public record ReissueJwtResult(
 	String accessToken,
