+++ Moyoy-Api/src/main/java/com/moyoy/api/auth/jwt/presentation/response/JwtReissueResponse.java
@@ -1,4 +1,4 @@
-package com.moyoy.api.auth.jwt.presentation;
+package com.moyoy.api.auth.jwt.presentation.response;
 
 public record JwtReissueResponse(
 	String accessToken) {
