+++ Moyoy-Api/src/test/java/com/moyoy/common/annotation/ControllerTest.java
@@ -1,0 +1,35 @@
+package com.moyoy.common.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
+import org.springframework.context.annotation.Import;
+import org.springframework.restdocs.RestDocumentationExtension;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import com.moyoy.api.common.ApiControllerAdvice;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@WebMvcTest(excludeFilters = {
+	@ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = OncePerRequestFilter.class)
+})
+@Import(ApiControllerAdvice.class)
+@AutoConfigureRestDocs
+@AutoConfigureMockMvc(addFilters = false)
+@ExtendWith(RestDocumentationExtension.class)
+public @interface ControllerTest{
+
+	/**
+	 * WebMvcTest 에서 테스트할 Controller 클래스 지정
+	 */
+	Class<?>[] controllers() default {};
+}
