--- Moyoy-Api/src/test/resources/application.yml
@@ -1,39 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:h2:mem:testdb;MODE=MYSQL;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false
-    driver-class-name: org.h2.Driver
-    username: sa
-    password:
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    properties:
-      hibernate:
-        format_sql: true
-    show-sql: true
-
-  flyway:
-    enabled: false
-
-  security:
-    oauth2:
-      client:
-        registration:
-          github:
-            client-id: ${CLIENT_ID}
-            client-secret: ${CLIENT_SECRET}
-            redirect-uri: ${REDIRECT_URI}
-            scope: ${SCOPE}
-
-  jwt:
-    secret: ${JWT_SECRET_KEY:test-secret-test-secret-test-secret-test-secret}
-
-  cors:
-    allow-origin: ${ALLOW_ORIGIN:http://localhost:5173}
-
-  cookie:
-    samesite: ${COOKIE_SAME_SITE:Strict}
-    domain: ${COOKIE_DOMAIN:localhost}
-
-  login:
-    default-uri: ${DEFAULT_URI:http://localhost:5173/menu}
