+++ Moyoy-Infra/build.gradle
@@ -14,7 +14,7 @@ dependencies {
     // MySQL
     runtimeOnly 'com.mysql:mysql-connector-j'
 
-    // H2
+    // Test
     testImplementation 'com.h2database:h2'
 
     // JPA & QueryDSL
