+++ Moyoy-Infra/src/main/resources/infra-dev.yml
@@ -1,12 +1,23 @@
 spring:
   config:
     import:
-      - "aws-parameterstore:/config/moyoy-test/" 
+      - "aws-parameterstore:/config/moyoy-test/"
 
   flyway:
     enabled: true
     locations: classpath:db/migration/{vendor}
     baseline-on-migrate: true
+
+  data:
+    redis:
+      host: ${REDIS_HOST:localhost}
+      port: ${REDIS_PORT:6379}
+
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: ${DB_URL}
+    username: ${DB_USERNAME}
+    password: ${DB_PASSWORD}
 
   jpa:
     hibernate:
@@ -17,3 +28,4 @@ aws:
     enabled: true
     prefix: /config
     name: moyoy-test
+
