+++ Moyoy-Infra/src/test/resources/application.yml
@@ -1,9 +1,11 @@
+## 인프라 모듈 테스트 용
 spring:
   datasource:
     url: jdbc:h2:mem:testdb;MODE=MYSQL;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false
     driver-class-name: org.h2.Driver
     username: sa
     password:
+
   jpa:
     hibernate:
       ddl-auto: create-drop
